+++ Modules/core/Controller/CoretilesController.php
@@ -130,7 +130,19 @@ class CoretilesController extends CorecookiesecureController {
 
             $allSpaces = $spaceModel->getSpaces('id');
             $spaceMap = [];
+            $userSpacesOnly = $this->request->getParameterNoException('mine') === '1' ? true: false;
+            if($userSpacesOnly) {
+                $spaces= [];
+            }
+            $userSpacesList = [];
+
             foreach($allSpaces as $space) {
+                if($logged && isset($_SESSION['login']) && $userSpacesOnly) {
+                    $isMemberOfSpace = (in_array($space["id"], $userSpaces['userSpaceIds'])) ? true : false;
+                    if($isMemberOfSpace) {
+                        $userSpacesList[] = $space;
+                    }
+                }
                 if ($logged && !in_array($space["id"], $userSpaces['spacesUserIsAdminOf']) && isset($_SESSION["login"])) {
                     if (!in_array($space["id"], $userSpaces['userPendingSpaceIds'])) {
                         $isMemberOfSpace = (in_array($space["id"], $userSpaces['userSpaceIds'])) ? true : false;
@@ -166,6 +178,7 @@ class CoretilesController extends CorecookiesecureController {
                 'content' => $content,
                 'spaces' => $spaces,
                 'spaceMap' => $spaceMap,
+                'userSpaces' => $userSpacesList,
                 'catalog' => $catalog,
                 'resources' => $resources,
                 'mainSubMenus' => [],